refactor: moved media parsers into folder + updated namespace
