+++ src/Http/ServerRequestBuilder.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 namespace BitFrame\Http;
 
 use BitFrame\Http\Normalizer\UploadedFilesNormalizer;
-use BitFrame\Parser\MediaParserNegotiator;
+use BitFrame\Parser\Media\MediaParserNegotiator;
 use BitFrame\Parser\ServerRequest\CookieHeaderParser;
 use BitFrame\Parser\ServerRequest\UriParser;
 use Psr\Http\Message\{
