+++ src/Parser/Media/JsonMediaParser.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace BitFrame\Parser;
+namespace BitFrame\Parser\Media;
 
 use function is_array;
 use function json_decode;
