+++ src/Parser/Media/XmlMediaParser.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace BitFrame\Parser;
+namespace BitFrame\Parser\Media;
 
 use SimpleXMLElement;
